Reduce file size with higher quality.
`ffmpeg` does a bad job while converting to GIF, so we convert to PNG.
We reduce file size by disabling dither and fuzzing 1% of the colors. Also,
the `-c` option is added to control the number of colors in the output
palette. Finally, because `gifsicle` has a better LWZ compression
optimization than `convert`, we finalize the size reducing with
`gifsicle`.

Also, we have added the `-d` option to control the delay between to
frame (we can find the balance between frame rate and delay to still
reduce the file size).

Note: on my example, 254Kb to 119Kb with an equivalent quality to the
video (before, I got white dots everywhere)!